refactor(improve): move preparation stage to improve/preparation.ts (D1c step A)
Move runImprovePreparationStage (1,657 LOC), runConsolidationPass (297), and
maybeAutoTuneThreshold out of improve.ts into improve/preparation.ts. improve.ts
5,289 -> 2,749 LOC (the orchestrator is no longer the god-module). One-directional
dependency: improve.ts imports the three functions; preparation.ts type-imports
the shared result types (ImprovePreparationResult/ImproveScope/ConsolidationPassResult/
AkmImproveOptions) — no runtime cycle. maybeAutoTuneThreshold re-exported for tests.
Verbatim move; lint 0 / tsc 0 / unit 5274-0.

Next (step B): decompose runImprovePreparationStage's ~15 phases into named functions.
